Move swaggerOptions.js to "routes" folder
